Improve configuring configurables actions
